Fix dry deposition seasonality
In chem/module_dep_simple.F, forces iseason to category 4 (winter, snow on
ground and subfreezing) if snow height >= 10 cm or if the land type is ice.
Also includes a case for the Southern Hemisphere, previously the seasonality
of dry deposition had a hard-coded asumption that the model was run in the
Northern Hemisphere for all mechanisms other than MOZART-based ones.