ADD VOLATILIZATION OF PARTICLE-BOUND MERCURY TO GAS-PHASE MERCURY
This updates the module_mosaic_set_heterogeneous.F code to include volatilization
of particle-bound mercury (HgP) to gas-phase mercury (HgX). This unspeciated mercury
is treated as HgCl2 which does not photolyze at tropospheric wavelengths. The volatilization
rate follows the empirical equilibrium of Amos et al. (2012).
Update to the registry to include the heterogeneous reaction rates.