UPDATE HG0 RE-EMISSION FROM SNOW AND SEA ICE
Updates to module_airsnow_emissions.F for Hg0 re-emission from snow and
sea ice. Hg0 re-emission is now calculated based on total deposited Hg2 to snow
and sea ice, assuming a net reduction rate constant for snow and ice surfaces.
It is also assumed that 60% of total Hg2 in snow is available for conversion to
Hg0 and re-emission.