Prevent CLDFRA overwrite in mixactivate
In phys/module_mixactivate.F, remove intent(inout) declarations for
CLDFRA and define a local cldfra_new array. This is to prevent the call
to cal_cldfra2 in mixactivate from overwriting the value of CLDFRA from
the radiation driver (which can be calculated by different methods than
cal_cldfra2 depending on icloud).