Add blowing snow aerosol emissions to WRF-Chem
Add blowing snow emissions of sea salt aerosols over sea ice,
compatible with the MOSAIC-4bin, MOSAIC-8bin, and GOCART aerosol
schemes.
Emissions are based on the Yang 2008 scheme. The values of the salinity
parameter (0.1 PSU) and of seasalt aerosol fractioning (N=5 particles
from 1 blowing snow particle) are from Huang and Jaegle (2017) instead
of the original Yang et al (2008) values.  The size distribution of the
emissions was integrated offline over the sizes of the MOSAIC and GOCART
bins.